Split server code to two files one for source and one for sink.
